update training script

+++ test/lmp/util/model/conftest.py
@@ -4,34 +4,77 @@ import os
 
 import pytest
 
-import lmp
+import lmp.util.path
 from lmp.model import BaseModel, ElmanNet
 from lmp.tknzr import BaseTknzr, CharTknzr
 
 
+@pytest.fixture(params=[1, 2])
+def d_blk(request) -> int:
+  """Dimension of each memory cell block."""
+  return request.param
+
+
+@pytest.fixture(params=[1, 2])
+def d_emb(request) -> int:
+  """Embedding dimension."""
+  return request.param
+
+
+@pytest.fixture(params=[1, 2])
+def d_hid(request) -> int:
+  """Hidden dimension."""
+  return request.param
+
+
+@pytest.fixture(params=[False, True])
+def is_uncased(request) -> bool:
+  """Respect cases if set to ``False``."""
+  return request.param
+
+
+@pytest.fixture(params=[-1, 10])
+def max_vocab(request) -> int:
+  """Maximum vocabulary size."""
+  return request.param
+
+
+@pytest.fixture(params=[0, 10])
+def min_count(request) -> int:
+  """Minimum token occurrence counts."""
+  return request.param
+
+
+@pytest.fixture(params=[1, 2])
+def n_blk(request) -> int:
+  """Number of memory cell blocks."""
+  return request.param
+
+
+@pytest.fixture(params=[0.1, 0.5])
+def p_emb(request) -> float:
+  """Embedding dropout probability."""
+  return request.param
+
+
+@pytest.fixture(params=[0.1, 0.5])
+def p_hid(request) -> float:
+  """Hidden units dropout probability."""
+  return request.param
+
+
 @pytest.fixture
-def tknzr() -> BaseTknzr:
+def tknzr(is_uncased: bool, max_vocab: int, min_count: int) -> BaseTknzr:
   """Tokenizer example."""
-  return CharTknzr(
-    is_uncased=True,
-    max_vocab=-1,
-    min_count=1,
-    tk2id={
-      CharTknzr.bos_tk: CharTknzr.bos_tkid,
-      CharTknzr.eos_tk: CharTknzr.eos_tkid,
-      CharTknzr.pad_tk: CharTknzr.pad_tkid,
-      CharTknzr.unk_tk: CharTknzr.unk_tkid,
-      'a': max(CharTknzr.bos_tkid, CharTknzr.eos_tkid, CharTknzr.pad_tkid, CharTknzr.unk_tkid) + 1,
-      'b': max(CharTknzr.bos_tkid, CharTknzr.eos_tkid, CharTknzr.pad_tkid, CharTknzr.unk_tkid) + 2,
-      'c': max(CharTknzr.bos_tkid, CharTknzr.eos_tkid, CharTknzr.pad_tkid, CharTknzr.unk_tkid) + 3,
-    },
-  )
+  tknzr = CharTknzr(is_uncased=is_uncased, max_vocab=max_vocab, min_count=min_count)
+  tknzr.build_vocab(batch_txt=['a', 'b', 'c'])
+  return tknzr
 
 
 @pytest.fixture
-def model(tknzr: BaseTknzr) -> BaseModel:
+def model(d_emb: int, d_hid: int, p_emb: float, p_hid: float, tknzr: BaseTknzr) -> BaseModel:
   """Save model example."""
-  return ElmanNet(d_emb=10, tknzr=tknzr)
+  return ElmanNet(d_emb=d_emb, d_hid=d_hid, p_emb=p_emb, p_hid=p_hid, tknzr=tknzr)
 
 
 @pytest.fixture
@@ -42,7 +85,9 @@ def ckpt_dir_path(exp_name: str, request) -> str:
   def fin() -> None:
     for file_name in os.listdir(abs_dir_path):
       os.remove(os.path.join(abs_dir_path, file_name))
-    os.removedirs(abs_dir_path)
+
+    if os.path.exists(abs_dir_path) and not os.listdir(abs_dir_path):
+      os.removedirs(abs_dir_path)
 
   request.addfinalizer(fin)
   return abs_dir_path
